+++ R/test-example.R
@@ -7,10 +7,33 @@
 #' @param path For `test_examples()`, path to directory containing Rd files.
 #'   For `test_example()`, path to a single Rd file. Remember the working
 #'   directory for tests is `tests/testthat`.
+#' @param rd A parsed Rd object, obtained from [tools::Rd_db()] or otherwise.
 #' @export
-test_examples <- function(path = "../../man") {
-  man <- dir(path, "\\.Rd$", full.names = TRUE)
-  lapply(man, test_example)
+test_examples <- function(path = "../..") {
+  res <- test_examples_source(path) %||% test_examples_installed(env_test$package)
+  if (is.null(res)) {
+    stop("Could not find examples", call. = FALSE)
+  }
+  res
+}
+
+test_examples_source <- function(path = "../..") {
+  if (!dir.exists(file.path(path, "man"))) {
+    return()
+  }
+  Rd <- tools::Rd_db(dir = path)
+  if (length(Rd) == 0) {
+    return()
+  }
+  lapply(Rd, test_rd)
+}
+
+test_examples_installed <- function(package = env_test$package) {
+  Rd <- tools::Rd_db(package = package)
+  if (length(Rd) == 0) {
+    return()
+  }
+  lapply(Rd, test_rd)
 }
 
 #' @export
@@ -22,9 +45,24 @@ test_example <- function(path) {
 
   env <- new.env(parent = globalenv())
 
-  ok <- test_code(path, parse(ex_path), env = globalenv())
+  ok <- test_code(path, parse(ex_path), env = env)
   if (ok) succeed(path)
 
   invisible()
 }
 
+#' @export
+#' @rdname test_examples
+test_rd <- function(rd) {
+  path <- attr(rd, "Rdfile")
+  ex_path <- file.path(tempdir(), paste0(tools::file_path_sans_ext(basename(path)), ".R"))
+  tools::Rd2ex(rd, ex_path)
+  if (!file.exists(ex_path)) return()
+
+  env <- new.env(parent = globalenv())
+
+  ok <- test_code(path, parse(ex_path), env = env)
+  if (ok) succeed(path)
+
+  invisible()
+}