+++ tests/testthat/test-expect-length.R
@@ -1,12 +1,16 @@
 context("expect_length")
 
-test_that("fails if not a vector", {
-  expect_failure(expect_length(environment(), 1), "not a vector")
-})
-
 test_that("length computed correctly", {
   expect_success(expect_length(1, 1))
-  expect_failure(expect_length(1, 2))
+  expect_failure(expect_length(1, 2), "has length 1, not length 2.")
+  expect_success(expect_length(1:10, 10))
+  expect_success(expect_length(letters[1:5], 5))
+})
+
+test_that("uses S4 length method", {
+  A <- setClass("ExpectLengthA", slots=c(x="numeric", y="numeric"))
+  setMethod("length", "ExpectLengthA", function (x) 5L)
+  expect_success(expect_length(A(x=1:9, y=3), 5))
 })
 
 test_that("returns input", {